fixed keys of maps
