--- src/Components/VirtualWorkshop/virtualWorkshop.css
@@ -1,3 +1,0 @@
-img{
-    max-height: 500px   
-}
