+++ src/Components/Workshop/PresencialWorkshop/PresencialWorkshop.jsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from "react";
 import { Helmet } from "react-helmet";
+import "./presencialWorkshop.css";
 import {
   Container,
   Col,
@@ -12,17 +13,16 @@ import {
   InputGroup,
   Carousel,
 } from "react-bootstrap";
-import "./virtualWorkshop.css";
-import CoursesDataServices from "../../Services/CoursesServices";
+import CoursesDataServices from "../../../Services/CoursesServices";
 import { useParams } from "react-router-dom";
-import { TGVimages, activity } from "../../assets/images";
+import { TGPimages, activity } from "../../../assets/images";
 import moment from "moment";
 
-const CursosVirtuales = () => {
+const CursosPresenciales = () => {
   const [index, setIndex] = useState(0);
+  const { name } = useParams();
   const [courses, setCourse] = useState([]);
 
-  // Función del Carousel de bootstrap
   const handleSelect = (selectedIndex, e) => {
     setIndex(selectedIndex);
   };
@@ -62,31 +62,34 @@ const CursosVirtuales = () => {
   return (
     <>
       <Helmet>
-        <title>Silvia Iurman - Talleres Virtuales</title>
+        <title>Silvia Iurman - Talleres Presenciales</title>
         <meta
           name="description"
-          content="Talleres virtuales de Silvia Iurman - Nuevas Constelaciones Familiares Argentina"
+          content="Talleres presenciales de Silvia Iurman - Nuevas Constelaciones Familiares Argentina"
         />
       </Helmet>
       <Container fluid>
-        <h1>Talleres Virtuales</h1>
+        <h1>Talleres presenciales</h1>
       </Container>
       {courses.map(
         (course) =>
-          course._id === "63d2d475dc2d95cfd1095e83" && (
-            <Container>
+          course._id === "63d2d339dc2d95cfd1095bdf" && (
+            <Container key={course._id}>
               <Row>
                 <Col lg={6} sm={12}>
                   <Carousel activeIndex={index} onSelect={handleSelect}>
-                    {TGVimages.map((image, index) => (
+                    {
+                      TGPimages.map((image, index) => (
                       <Carousel.Item key={index}>
                         <img
-                          className="d-block w-100"
+                          className="d-block w-100 h-100 object-fit-cover"
                           src={image.src}
                           alt={image.alt}
+                          loading="lazy"
                         />
                       </Carousel.Item>
-                    ))}
+                    ))
+                    }
                   </Carousel>
                   {/* <Image className="img-fluid" src={activity} alt="" /> */}
                 </Col>
@@ -247,4 +250,4 @@ const CursosVirtuales = () => {
   );
 };
 
-export default CursosVirtuales;
+export default CursosPresenciales;
